% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_references.R
\name{add_function_references}
\alias{add_function_references}
\title{Add a Markdown file with function references}
\usage{
add_function_references(include_internal = TRUE)
}
\arguments{
\item{include_internal}{Boolean indicating if you want to include the documentation of internal (i.e non-exported functions). Default is TRUE. See Details.}
}
\value{
A Markdown file called "func_reference.md" in the folder "docs".
}
\description{
Add a Markdown file with function references
}
\details{
This function is automatically called with \code{use_docsify()} by default. However, if you didn't want to create it at the beginning but you changed your mind after having run \code{use_docsify()}, you can run it on its own.

If you don't want to include internal functions (i.e functions that are not exported by the package), include "@keywords internal" in the roxygen block of the function concerned, and use \code{include_internal = FALSE}.
}
