% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_md_rmd.R
\name{add_md_rmd}
\alias{add_md_rmd}
\alias{add_md}
\alias{add_rmd}
\title{Create a Markdown (or R Markdown) file to populate your documentation.}
\usage{
add_md(name, open = TRUE)

add_rmd(name, open = TRUE)
}
\arguments{
\item{name}{Name of the .md (or .Rmd) file to create. If the file already exists, it will return an error.}

\item{open}{Open or not the files created. Default is TRUE.}
}
\value{
Creates a Markdown file in "docs".
}
\description{
Create a Markdown (or R Markdown) file to populate your documentation.
}
\examples{
\dontrun{

library(docsifier)

# Create a test folder and a test package for the example

test_folder <- tempdir()
setwd(test_folder)
devtools::create("dummy")
setwd("dummy")

# Generate the minimal documentation for docsify.js

use_docsify()

# Create a new .md in "/docs"

add_md("test")

# Will output an error because "test.md" already exists

add_md("test")
}
}
