% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_sidebar.R
\name{add_sections}
\alias{add_sections}
\alias{add_code_of_conduct}
\alias{add_news}
\alias{add_license}
\title{Add the Code of Conduct, the License and the NEWS file as section or subsection}
\usage{
add_code_of_conduct(section_above = NULL, type = "section")

add_news(section_above = NULL, type = "section")

add_license(section_above = NULL, type = "section")
}
\arguments{
\item{section_above}{Name of the section under which the new section/subsection will be placed. If NULL (default), it will be placed under all the other sections.}

\item{type}{"section" or "subsection"}
}
\description{
Add the Code of Conduct, the License and the NEWS file as section or subsection
}
