% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_sidebar.R
\name{add_to_sidebar}
\alias{add_to_sidebar}
\title{Add a section to _sidebar.md}
\usage{
add_to_sidebar(section_name, subsection_name = NULL, subsection_md_file = NULL)
}
\arguments{
\item{section_name}{Name of the section created with this file. If NULL, the name of the file is used.}

\item{filename}{Name of the file to add, must be a Markdown (.md) file}
}
\value{
Adds a section below all the other sections in _sidebar.md
}
\description{
Add a section to _sidebar.md
}
\keyword{internal}
