% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_vignettes.R
\name{add_vignettes}
\alias{add_vignettes}
\title{Convert, move and insert vignettes}
\usage{
add_vignettes(section_name = "Articles", section_above = "Home")
}
\arguments{
\item{section_name}{Name of the section containing the vignettes. Default is "Articles".}

\item{section_above}{Section below which the section "Articles" (or the name given in \code{section_name}) will be placed. By default, it will be placed just under "Home".}
}
\description{
Convert, move and insert vignettes
}
