% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_snake_case}
\alias{as_snake_case}
\title{Transform string in snake case}
\usage{
as_snake_case(x)
}
\arguments{
\item{x}{String to transform}
}
\description{
Transform string in snake case
}
\keyword{internal}
