% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_references.R
\name{build_function_reference}
\alias{build_function_reference}
\title{Gather the important info in .Rd files}
\usage{
build_function_reference(include_internal = TRUE)
}
\description{
Gather the important info in .Rd files
}
\keyword{internal}
