% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_in_index.R
\name{change_in_index}
\alias{change_in_index}
\alias{change_github_in_index}
\alias{change_pkgname_in_index}
\title{Change fields in index.html}
\usage{
change_github_in_index()

change_pkgname_in_index()
}
\description{
Change fields in index.html
}
\keyword{(title)}
\keyword{Function}
\keyword{as}
\keyword{automatically}
\keyword{changes}
\keyword{documentation}
\keyword{github}
\keyword{in}
\keyword{index.html}
\keyword{internal}
\keyword{name}
\keyword{of}
\keyword{package}
\keyword{puts}
\keyword{that}
\keyword{the}
\keyword{url}
