% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{folder_is_empty}
\alias{folder_is_empty}
\title{Detect if the folder is empty}
\usage{
folder_is_empty(x)
}
\arguments{
\item{x}{Name of the folder}
}
\description{
Detect if the folder is empty
}
\keyword{internal}
