% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_img_paths}
\alias{get_img_paths}
\title{Find image paths in Markdown files}
\usage{
get_img_paths(filename)
}
\arguments{
\item{filename}{Name of the Markdown file from we which extract the image paths}
}
\description{
Find image paths in Markdown files
}
\keyword{internal}
