% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_in_text}
\alias{get_in_text}
\title{Obtain words between curly braces in .Rd files}
\usage{
get_in_text(item, text)
}
\arguments{
\item{item}{Item to extract from the .Rd file. Can be any character vector among those between "\" and "\{" in a .Rd file. Use \code{NULL} to return all items between curly braces.}

\item{text}{Text from which to extract.}
}
\description{
Obtain words between curly braces in .Rd files
}
\keyword{internal}
