% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_in_index.R
\name{github_url_in}
\alias{github_url_in}
\title{Extract github URL from URL or BugReports fields in DESCRIPTION}
\usage{
github_url_in(field = "URL")
}
\arguments{
\item{field}{URL or BugReports (or another field where the URL could appear)}
}
\description{
Extract github URL from URL or BugReports fields in DESCRIPTION
}
\keyword{internal}
