% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_after}
\alias{insert_after}
\title{Insert a line in a text file}
\usage{
insert_after(file, where, insert)
}
\arguments{
\item{file}{File in which insert a line}

\item{where}{After which line we want to insert some text}

\item{insert}{Text to insert}
}
\description{
Insert a line in a text file
}
\keyword{internal}
