% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_it_a_package}
\alias{is_it_a_package}
\title{Detect if we are in a package environment or not}
\usage{
is_it_a_package()
}
\description{
Detect if we are in a package environment or not
}
\keyword{internal}
