% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_quoted}
\alias{is_quoted}
\title{Detect if a string is between quotation marks}
\usage{
is_quoted(string)
}
\arguments{
\item{string}{String concerned}
}
\description{
Detect if a string is between quotation marks
}
\keyword{internal}
