% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_sidebar.R
\name{link_md_to_section}
\alias{link_md_to_section}
\title{Link a .md file to a section in _sidebar.md}
\usage{
link_md_to_section(md_file, section_name)
}
\arguments{
\item{md_file}{Name of the .md file to link}

\item{section_name}{Name of the section to which link the .md file}
}
\description{
Link a .md file to a section in _sidebar.md
}
