% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{message_validate}
\alias{message_validate}
\title{Wrappers for cli messages}
\usage{
message_validate(x)
}
\arguments{
\item{x}{Message (character vector)}
}
\description{
Wrappers for cli messages
}
\keyword{internal}
