% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_available.R
\name{node_npm}
\alias{node_npm}
\alias{npm_available}
\alias{node_available}
\title{Check if Node and npm are installed}
\usage{
npm_available()

node_available()
}
\description{
Check if Node and npm are installed
}
