% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_docsify.R
\name{preview_docsify}
\alias{preview_docsify}
\title{Preview the documentation in a webpage or in viewer}
\usage{
preview_docsify()
}
\value{
Runs a preview of the site generated by docsify.js in RStudio viewer.
}
\description{
This runs a preview of the site generated by docsify.js in RStudio viewer. This function must be called after having created the essential files with \code{use_docsify()}.
}
\examples{
\dontrun{
# Create a test folder and a test package for the example

test_folder <- tempdir()
setwd(test_folder)
devtools::create("dummy")
setwd("dummy")

# Generate the minimal documentation for docsify.js

use_docsify()

# Run the preview

preview_docsify()
}
}
