% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_vignettes.R
\name{put_vignettes_in_sidebar}
\alias{put_vignettes_in_sidebar}
\title{Insert the vignettes in sidebar}
\usage{
put_vignettes_in_sidebar(section_name = "Articles", section_above = "Home")
}
\arguments{
\item{section_name}{Name of the section containing the vignettes. Default is "Articles".}

\item{section_above}{Section below which the section "Articles" (or the name given in \code{section_name}) will be placed. By default, it will be placed just under "Home".}
}
\value{
Modifies "_sidebar.md" to insert the section containing the vignettes. Each vignette is a subsection.
}
\description{
This function creates the section "Articles" (or with another name) and puts all vignettes in "docs/articles" as subsections.
}
\keyword{internal}
