% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_sidebar.R
\name{readme_as_homepage}
\alias{readme_as_homepage}
\title{Put README as homepage}
\usage{
readme_as_homepage()
}
\description{
This function simply moves README.md in the docs and rename it "homepage.md". This is the only thing to do because homepage.md is defined as the .md file associated to "Home" in index.html.
}
