% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_img_paths}
\alias{replace_img_paths}
\title{Replace old image paths by new ones}
\usage{
replace_img_paths(filename)
}
\arguments{
\item{filename}{Name of the Markdown file in which we replace the image paths}
}
\description{
Replace old image paths by new ones
}
\keyword{internal}
