% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unquote_title}
\alias{unquote_title}
\title{Unquote the title given in .Rmd file's YAML}
\usage{
unquote_title(filename)
}
\arguments{
\item{filename}{File concerned}
}
\description{
Unquote the title given in .Rmd file's YAML
}
\keyword{internal}
