% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_docsify}
\alias{update_docsify}
\title{Update files in docs}
\usage{
update_docsify(include_internal = FALSE)
}
\arguments{
\item{include_internal}{Boolean indicating if internal functions should be included in the "Reference" page.}
}
\description{
Update files in docs
}
