% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_file}
\alias{update_file}
\title{Update a file in folder "docs"}
\usage{
update_file(filename, name_in_doc = NULL)
}
\arguments{
\item{filename}{Name of the file which is up-to-date.}

\item{name_in_doc}{Name of the file to be updated in "docs".}
}
\value{
If file doesn't exist in "docs" then it is created. If file already exists in "docs", then it will overwrite it with the new version.
}
\description{
Update a file in folder "docs"
}
\keyword{internal}
