% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vignettes_differ}
\alias{vignettes_differ}
\title{Check if vignettes in folder "vignettes" and in folder "docs/articles" differ}
\usage{
vignettes_differ(x, y)
}
\arguments{
\item{x, y}{Names of the two vignettes to compare}
}
\value{
Boolean
}
\description{
Since the output of the vignette in the folder "vignette" is "html_vignette" and the output of the vignette in the folder "docs/articles" is "github_document", there will necessarily be changes. Therefore, the comparison is made on the files without the YAML.
}
\keyword{internal}
