% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_css}
\alias{add_css}
\title{Add a CSS file to your documentation}
\usage{
add_css(name = NULL, open = TRUE)
}
\arguments{
\item{name}{Name to give to the CSS file you want to create. If \code{NULL}, the file will be named \code{custom.css}}

\item{open}{Open or not the file created. Default is TRUE.}
}
\value{
Creates a CSS file in "docs/_assets/css"
}
\description{
Add a CSS file to your documentation
}
\examples{
\dontrun{

library(docsifier)

# Create a test folder and a test package for the example

test_folder <- tempdir()
setwd(test_folder)
devtools::create("dummy")
setwd("dummy")

# Generate the minimal documentation for docsify.js

init_docsify()

# Create "custom.css"  in "/docs/_assets/css"

add_css()
}
}
