% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_references.R
\name{add_function_references}
\alias{add_function_references}
\title{Add a Markdown file with function references}
\usage{
add_function_references(include_internal = TRUE)
}
\arguments{
\item{include_internal}{Boolean indicating if you want to include the documentation of internal (i.e non-exported functions). Default is TRUE. See Details.}
}
\value{
Creates a Markdown file called "func_reference.md" in the folder "docs", and edit "_sidebar.md" to add a "Reference" section.
}
\description{
Add a Markdown file with function references
}
\details{
This function is automatically called with \code{use_docsify()} by default. However, if you didn't want to create it at the beginning but you changed your mind after having run \code{use_docsify()}, you can run it on its own.

If you don't want to include internal functions (i.e functions that are not exported by the package), include "@keywords internal" in the roxygen block of the function concerned, and use \code{include_internal = FALSE}.
}
\examples{
\dontrun{

library(docsifier)

# Create a test folder and a test package for the example

test_folder <- tempdir()
setwd(test_folder)
devtools::create("dummy")
setwd("dummy")

# Generate the minimal documentation for docsify.js

use_docsify(add_reference = FALSE)

# Generate the "Reference" page in the documentation

add_function_references()
}
}
