% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_md.R
\name{add_md_rmd}
\alias{add_md_rmd}
\alias{add_md}
\alias{add_rmd}
\title{Create a Markdown (or R Markdown) file to populate your documentation.}
\usage{
add_md(name)

add_rmd(name)
}
\arguments{
\item{name}{Name of the .md (or .Rmd) file to create. If the file already exists, it will return an error.}
}
\description{
Create a Markdown (or R Markdown) file to populate your documentation.
}
\examples{
\dontrun{
# Create a new .md in "/docs"
add_md("test")

# Will output an error because "test.md" already exists
add_md("test")
}
}
