% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_sidebar.R
\name{add_to_sidebar}
\alias{add_to_sidebar}
\title{Add a section to _sidebar.md}
\usage{
add_to_sidebar(
  file,
  name,
  section_above = NULL,
  type = c("section", "subsection")
)
}
\arguments{
\item{file}{Name of the file to copy in "docs"}

\item{name}{Name of the section or subsection to add}

\item{section_above}{Name of the section under which the new section/subsection will be placed. If NULL (default), it will be placed under all the other sections.}

\item{type}{"section" or "subsection"}
}
\value{
Adds a section below all the other sections in _sidebar.md
}
\description{
Add a section to _sidebar.md
}
