% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_docsify.R
\name{use_docsify}
\alias{use_docsify}
\title{Create the structure to use docsify in an R package}
\usage{
use_docsify(open = TRUE)
}
\arguments{
\item{open}{Boolean indicating whether to open "index.html" and "homepage.md" following their creation.}
}
\description{
Create the structure to use docsify in an R package
}
\details{
This function creates the folder "/docs" if it doesn"t already exist, and "index.html" to create the documentation with docsify. The structure of "index.html" is automatically created and you can complete it with your custom info.

You can use \code{add_md()} to add Markdown files in "/docs" to populate the documentation.
}
